refactor(forge): split GitHub auth methods into github_auth.go (#1025)
Pure same-package code motion: move VerifyToken, TokenInfo, and
parseGitHubScopes out of github.go into a new sibling github_auth.go,
the GitHub token-verification / identity seam. No behaviour change;
go doc -u -all is byte-identical.
